requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pycryptodome==3.23.0
openai==1.109.1
